refactor(create): move movie create files around
