+++ lib/plugins/features/movies/index.js
@@ -1,7 +1,7 @@
 'use strict';
 
-const Controller     = require('./controller');
-const MovieValidator = require('../../../validators/movie');
+const Controller      = require('./controller');
+const CreateValidator = require('../../../validators/movies/create');
 
 exports.register = (server, options, next) => {
 
@@ -11,7 +11,7 @@ exports.register = (server, options, next) => {
     config: {
       handler: (request, reply) => reply(Controller.create(request.payload)),
       validate: {
-        payload: MovieValidator
+        payload: CreateValidator
       }
     }
   }]);
