+++ test/validators/movies/create.test.js
@@ -2,15 +2,15 @@
 
 const Joi = require('joi');
 
-const MovieValidator = require('../../lib/validators/movie');
+const MovieCreateValidator = require('../../../lib/validators/movies/create');
 
-describe('movie validator', () => {
+describe('movie create validator', () => {
 
   it('succeeds when all allowed params are included', () => {
     const name = 'title';
     const release_year = 2017;
     const payload = { name, release_year };
-    const result = Joi.validate(payload, MovieValidator);
+    const result = Joi.validate(payload, MovieCreateValidator);
 
     expect(result.error).to.be.null;
   });
@@ -19,7 +19,7 @@ describe('movie validator', () => {
 
     it('is required', () => {
       const payload = {};
-      const result = Joi.validate(payload, MovieValidator);
+      const result = Joi.validate(payload, MovieCreateValidator);
 
       expect(result.error.details[0].path).to.eql('name');
       expect(result.error.details[0].type).to.eql('any.required');
@@ -28,7 +28,7 @@ describe('movie validator', () => {
     it('is less than 255 characters', () => {
       const name = 'a'.repeat(256);
       const payload = { name };
-      const result = Joi.validate(payload, MovieValidator);
+      const result = Joi.validate(payload, MovieCreateValidator);
 
       expect(result.error.details[0].path).to.eql('name');
       expect(result.error.details[0].type).to.eql('string.max');
@@ -37,7 +37,7 @@ describe('movie validator', () => {
     it('is nonempty', () => {
       const name = '';
       const payload = { name };
-      const result = Joi.validate(payload, MovieValidator);
+      const result = Joi.validate(payload, MovieCreateValidator);
 
       expect(result.error.details[0].path).to.eql('name');
       expect(result.error.details[0].type).to.eql('any.empty');
@@ -51,7 +51,7 @@ describe('movie validator', () => {
       const name = 'Dom The Bomb';
       const release_year = 1877;
       const payload = { name, release_year };
-      const result = Joi.validate(payload, MovieValidator);
+      const result = Joi.validate(payload, MovieCreateValidator);
 
       expect(result.error.details[0].path).to.eql('release_year');
       expect(result.error.details[0].type).to.eql('number.min');
@@ -61,7 +61,7 @@ describe('movie validator', () => {
       const name = 'Dom The Bomb 2: Domlectric Boogaloo';
       const release_year = 10000;
       const payload = { name, release_year };
-      const result = Joi.validate(payload, MovieValidator);
+      const result = Joi.validate(payload, MovieCreateValidator);
 
       expect(result.error.details[0].path).to.eql('release_year');
       expect(result.error.details[0].type).to.eql('number.max');
@@ -70,7 +70,7 @@ describe('movie validator', () => {
     it('is optional', () => {
       const name = 'Dom 3 Bomb';
       const payload = { name };
-      const result = Joi.validate(payload, MovieValidator);
+      const result = Joi.validate(payload, MovieCreateValidator);
 
       expect(result.error).to.be.null;
     });
@@ -79,7 +79,7 @@ describe('movie validator', () => {
       const name = 'Dom Fast Dom 4ious';
       const release_year = 2017.89;
       const payload = { name, release_year };
-      const result = Joi.validate(payload, MovieValidator);
+      const result = Joi.validate(payload, MovieCreateValidator);
 
       expect(result.error.details[0].path).to.eql('release_year');
       expect(result.error.details[0].type).to.eql('number.integer');
